example-projects/crewai-medium-1/requirements.txt
crewai==0.11.0
google-auth>=2.16.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.80.0
openai>=1.0.0
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.28.0
langchain
langchain_community